check_models.py
import google.generativeai as genai
import os
import json

# Try to load credentials from token.json or similar to authenticate if needed, 
# but genai.list_models needs an API key. 
# I will assume the user has to input it in the UI, so I can't run this script 
# without the key.
# However, I can't ask the user for the key just to run this script.
# I will check if I can get the key from the environment or if I can mock it 
# (mocking won't work for hitting the API).
# Wait, the app asks for the key: api_key = st.text_input(...)
# So I cannot run this script standalone unless I have the key.

# I will assume "Gemini 3 Flash" might not exist and the user implies "Gemini 1.5 Flash" 
# or the latest. I'll code the app to find 'flash' and verify in the UI code.
# But better: I will update the app to PRINT the models to the console or logs 
# so I can see them if I were running it, but I can't see the running app logs.

# Alternative: I'll assume 1.5 Flash is safe, or 2.0. 
# "Gemini 3" is likely a confusion by the user or a very new release.
# I will search specifically for "gemini-3.0-flash" in the list inside the app logic.

print("Script skipped: Need API Key")
