requirements.txt
streamlit
pandas
google-analytics-data
python-slugify
google-auth-oauthlib
google-api-python-client
google-auth
google-analytics-admin
google-generativeai